Use #split to separate party id from member name
Previously, we were using a regex but that failed to capture the
party name of one member. The regex was looking for uppercase
letters that were preceded by a '(' but the member had two sets
of letters wrapped in parentheses.

The party_id is now separated from the member name using #split.
The last element is used as the party_id is
always the last set of letters wrapped in brackets.