Combine fore and surnames with matai titles
This article refers to the member by her matai title: http://www.samoaobserver.ws/en/18_03_2016/local/3805/Fiame-named-Deputy-Prime-Minister-Fonotoe-Faumuina-out-of-Cabinet.htm

From this, it's clear are considered to be actual names rather than prefixes.

This commit combines the fore and surnames of the members with their matai
titles and returns the combined string in the `:name` field.